+++ app/__init__.py
@@ -1,174 +1,68 @@
-from flask import Flask, send_from_directory
-from .extensions import db, migrate, login_manager, babel_ext, limiter, csrf
-from .config import Config
-import cloudinary  # <-- เพิ่ม import
+# app/__init__.py
+# นี่คือ App Factory
+from flask import Flask, send_from_directory, jsonify
 
-from .utils.helpers import format_as_bangkok_time, from_json_string, get_anonymous_name
-from .forms.upload import EmptyForm
-from .forms.owner import ROOM_TYPE_CHOICES
+# 1. Import จาก app/core/
+from .core.config import Config
+from .core import extensions, dependencies, context_processors, cli
+from .utils.helpers import format_as_bangkok_time, from_json_string
 
+# 2. Import Blueprints
 from .blueprints.public import bp as public_bp
 from .blueprints.owner import bp as owner_bp
 from .blueprints.admin import bp as admin_bp
 from .blueprints.auth import bp as auth_bp
 from .blueprints.api import bp as api_bp
 
-from .repositories.sqlalchemy.user_repo_sql import SqlUserRepo
-from .repositories.sqlalchemy.property_repo_sql import SqlPropertyRepo
-from .repositories.sqlalchemy.approval_repo_sql import SqlApprovalRepo
-from .repositories.sqlalchemy.review_repo_sql import SqlReviewRepo
-from .repositories.sqlalchemy.review_report_repo_sql import SqlReviewReportRepo
-
-from .services.auth_service import AuthService
-from .services.property_service import PropertyService
-from .services.search_service import SearchService
-from .services.approval_service import ApprovalService
-from .services.upload_service import UploadService
-from .services.dashboard_service import DashboardService
-from .services.review_service import ReviewService
-from .services.review_management_service import ReviewManagementService
-from .services.history_service import HistoryService
-
-def register_dependencies(app: Flask):
-    container = {}
-    container["user_repo"] = SqlUserRepo()
-    container["property_repo"] = SqlPropertyRepo()
-    container["approval_repo"] = SqlApprovalRepo()
-    container["review_repo"] = SqlReviewRepo()
-    container["review_report_repo"] = SqlReviewReportRepo()
-    container["upload_service"] = UploadService(app.config.get("UPLOAD_FOLDER", "uploads"))
-    container["auth_service"] = AuthService(
-        user_repo=container["user_repo"],
-        upload_service=container["upload_service"]
-    )
-    container["property_service"] = PropertyService(container["property_repo"])
-    container["search_service"] = SearchService(container["property_repo"])
-    container["approval_service"] = ApprovalService(container["approval_repo"], container["property_repo"])
-    container["review_service"] = ReviewService(container["review_repo"])
-    container["review_management_service"] = ReviewManagementService(
-        review_repo=container["review_repo"],
-        report_repo=container["review_report_repo"],
-        prop_repo=container["property_repo"]
-    )
-    container["dashboard_service"] = DashboardService(
-        user_repo=container["user_repo"],
-        property_repo=container["property_repo"],
-        approval_repo=container["approval_repo"],
-        review_report_repo=container["review_report_repo"] 
-    )
-    container["history_service"] = HistoryService(container["property_repo"])
-    
-    if not hasattr(app, "extensions"):
-        app.extensions = {}
-    app.extensions["container"] = container
-
-def create_app() -> Flask:
+def create_app(config_class=Config) -> Flask:
+    """
+    ฟังก์ชัน Factory สำหรับสร้าง Flask Application Instance
+    """
     app = Flask(__name__, instance_relative_config=True)
-    app.config.from_object(Config)
+    
+    # 3. โหลดการตั้งค่า (Config)
+    app.config.from_object(config_class)
 
-    # --- vvv เพิ่มส่วนนี้เข้าไป vvv ---
-    cloudinary.config(
-        cloud_name=app.config.get("CLOUDINARY_CLOUD_NAME"),
-        api_key=app.config.get("CLOUDINARY_API_KEY"),
-        api_secret=app.config.get("CLOUDINARY_API_SECRET"),
-        secure=True
-    )
-    # --- ^^^ สิ้นสุดส่วนที่เพิ่ม ^^^ ---
+    # 4. Initialize Extensions (จาก app/core/extensions.py)
+    extensions.db.init_app(app)
+    extensions.migrate.init_app(app, extensions.db)
+    extensions.login_manager.init_app(app)
+    extensions.babel_ext.init_app(app)
+    extensions.limiter.init_app(app)
+    extensions.csrf.init_app(app)
 
-    db.init_app(app)
-    migrate.init_app(app, db)
-    login_manager.init_app(app)
-    babel_ext.init_app(app)
-    limiter.init_app(app)
-    csrf.init_app(app)
+    # 5. ลงทะเบียน Dependencies (จาก app/core/dependencies.py)
+    with app.app_context():
+        dependencies.register_dependencies(app)
 
+    # 6. ลงทะเบียน Jinja Filters
     app.jinja_env.filters['to_bkk_time'] = format_as_bangkok_time
     app.jinja_env.filters['fromjson'] = from_json_string
 
-    @app.context_processor
-    def inject_global_vars():
-        history_service = app.extensions["container"].get("history_service")
-        recently_viewed = []
-        if history_service:
-            recently_viewed = history_service.get_viewed_properties()
+    # 7. ลงทะเบียน Context Processors (จาก app/core/context_processors.py)
+    app.context_processor(context_processors.inject_global_vars)
 
-        room_type_map = dict(ROOM_TYPE_CHOICES)
-        
-        return dict(
-            empty_form=EmptyForm(),
-            ROOM_TYPES=room_type_map,
-            recently_viewed_properties=recently_viewed,
-            anonymous_name=get_anonymous_name() 
-        )
-
-    with app.app_context():
-        register_dependencies(app)
-
+    # 8. ลงทะเบียน Blueprints
     app.register_blueprint(public_bp)
-    app.register_blueprint(owner_bp, url_prefix="/owner")
-    app.register_blueprint(admin_bp, url_prefix="/admin")
-    app.register_blueprint(auth_bp, url_prefix="/auth")
-    app.register_blueprint(api_bp, url_prefix="/api")
+    app.register_blueprint(owner_bp) # url_prefix ถูกกำหนดใน __init__.py ของ owner
+    app.register_blueprint(admin_bp) # url_prefix ถูกกำหนดใน __init__.py ของ admin
+    app.register_blueprint(auth_bp)  # url_prefix ถูกกำหนดใน __init__.py ของ auth
+    app.register_blueprint(api_bp)   # url_prefix ถูกกำหนดใน __init__.py ของ api
 
-    # --- vvv ลบ route นี้ออกไป vvv ---
-    # @app.route('/uploads/<path:filename>')
-    # def serve_uploads(filename):
-    #     return send_from_directory(
-    #         app.config['UPLOAD_FOLDER'],
-    #         filename,
-    #         as_attachment=False
-    #     )
-    # --- ^^^ สิ้นสุดส่วนที่ลบ ^^^ ---
+    # 9. ลงทะเบียน CLI Commands (จาก app/core/cli.py)
+    cli.register_commands(app)
+
+    # 10. ลงทะเบียน Routes พื้นฐาน
+    @app.route('/uploads/<path:filename>')
+    def serve_uploads(filename):
+        return send_from_directory(
+            app.config['UPLOAD_FOLDER'],
+            filename,
+            as_attachment=False
+        )
 
     @app.get("/health")
     def health():
-        return {"ok": True}
-
-    @app.cli.command("seed_amenities")
-    def seed_amenities():
-        from app.models.property import Amenity
-        from app.extensions import db
-        data = [
-            ("pet","อนุญาตสัตว์เลี้ยง","Pets allowed"), ("ac","เครื่องปรับอากาศ","Air conditioning"),
-            ("guard","รปภ.","Security guard"), ("cctv","กล้อง CCTV","CCTV"),
-            ("fridge","ตู้เย็น","Refrigerator"), ("bed","เตียง","Bed"),
-            ("heater","เครื่องทำน้ำอุ่น","Water heater"), ("internet","อินเทอร์เน็ต","Internet"),
-            ("tv","ทีวี","TV"), ("sofa","โซฟา","Sofa"),
-            ("wardrobe","ตู้เสื้อผ้า","Wardrobe"), ("desk","โต๊ะทำงาน","Desk"),
-        ]
-        for code, th, en in data:
-            if not Amenity.query.filter_by(code=code).first():
-                db.session.add(Amenity(code=code, label_th=th, label_en=en))
-        db.session.commit()
-        print("Seeded amenities ✅")
-
-    @app.cli.command("seed_sample")
-    def seed_sample():
-        from app.models.user import Owner, Admin
-        from app.models.property import Property
-        from app.extensions import db
-        from werkzeug.security import generate_password_hash
-        
-        location_pin_data = {"type": "Point", "coordinates": [100.7758, 13.7292]}
-
-        if not Owner.query.filter_by(email="owner@example.com").first():
-            o = Owner(full_name_th="เจ้าของตัวอย่าง", citizen_id="1101700203451",
-                      email="owner@example.com", password_hash=generate_password_hash("password"),
-                      is_active=True, approval_status='approved') # Set owner as active
-            db.session.add(o)
-            db.session.commit()
-            
-            p = Property(owner_id=o.id, dorm_name="ตัวอย่างหอพัก", room_type="studio",
-                         rent_price=6500,
-                         location_pin=location_pin_data,
-                         workflow_status=Property.WORKFLOW_APPROVED)
-            db.session.add(p)
-            db.session.commit()
-            
-        if not Admin.query.filter_by(username="admin").first():
-            a = Admin(username="admin", password_hash=generate_password_hash("admin"), display_name="Administrator")
-            db.session.add(a)
-            db.session.commit()
-        print("Seeded sample data ✅")
+        return jsonify({"ok": True})
 
     return app
