+++ app/blueprints/api/__init__.py
@@ -1,3 +1,4 @@
+# app/blueprints/auth/__init__.py
 from flask import Blueprint
-bp = Blueprint("api", __name__)
-from . import routes  # noqa: E402,F401
+bp = Blueprint("auth", __name__, url_prefix="/auth") # <-- เพิ่ม url_prefix ที่นี่
+from . import routes  # <-- เพิ่มบรรทัดนี้
